% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esurge.R
\name{esrg_aictbl}
\alias{esrg_aictbl}
\title{\code{esrg_aictbl} creates AIC table of model results}
\usage{
esrg_aictbl(rname)
}
\arguments{
\item{rname}{name of results model summary file (eg., dipper.mod)}
}
\value{
returns AIC table.
}
\description{
\code{esrg_aictbl} creates AIC table of model results
}
\examples{
   esrg_aictbl(nstates=4, nevents=4)

  \dontrun{
   v=esrg_aictbl()
  }


}
\seealso{
\code{\link{esrg_run}},
\code{\link{esrg_open}}
}

