% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esurge.R
\name{esrg_close}
\alias{esrg_close}
\title{\code{esrg_close} closes E-SURGE window}
\usage{
esrg_close()
}
\value{
returns nothing.
}
\description{
\code{esrg_close} closes E-SURGE window
}
\examples{
   rslt='dipper.mod'; data='ed.rh';
   ngrps=1; nicov=0; nstates=2; nevents=2; nage=1
   v=esrg_open(rslt,data,ngrps,nicov,nstates,nevents,nage,'noinit')
   v=esrg_run(rslt,'i','t','t')     #   run model phi(t)p(t)    (init is ignored for this model.)
   esrg_close()

  \dontrun{
   v=esrg_close()
  }

}
\seealso{
\code{\link{esrg_run}},
\code{\link{esrg_open}}
}

