% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esurge.R
\name{esrg_make_gepat_trans}
\alias{esrg_make_gepat_trans}
\title{\code{esrg_make_gepat_trans} creates GEPAT pattern file for 2-step transitions}
\usage{
esrg_make_gepat_trans(nstates, nevents, steps = 2, JMV = F)
}
\arguments{
\item{nstates}{number of states}

\item{nevents}{number of events}

\item{steps}{number of steps for transition matrix (default=2)}

\item{JMV}{=T if JMV model (p(from.to.?)), =F if CAS model}
}
\value{
returns nothing, but saves patterns to file, trans.pat and prints
}
\description{
\code{esrg_make_gepat_trans} creates GEPAT pattern file for 2-step transitions
}
\examples{
   esrg_make_gepat_trans(nstates=4, nevents=4)

  \dontrun{
   v=esrg_make_gepat_trans(4,4)
  }

}
\seealso{
\code{\link{esrg_run}},
\code{\link{esrg_open}}
}

