% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esurge.R
\name{esrg_open}
\alias{esrg_open}
\title{\code{esrg_open} opens E-SURGE, reads data}
\usage{
esrg_open(rslt, data, ngrps, nicov, nstates, nevents, nage, factopt = "",
  slow = NA, dlay = NA, verb = F, esrg = "")
}
\arguments{
\item{rslt}{filename for result AIC table}

\item{data}{filename for input data file}

\item{ngrps}{number of groups in data}

\item{nicov}{number of individual covariates in data}

\item{nstates}{number of states in data}

\item{nevents}{number of events in data}

\item{nage}{number of age-classes in data}

\item{factopt}{factorization option (="init" for init/trans/event, "noinit" for trans/event)}

\item{slow}{determines how fast characters are sent to E-SURGE (eg. slow=500 -> 1/2 sec per char)}

\item{dlay}{time between send actions (eg., 500 -> 1/2 sec delay between each send)}

\item{verb}{verbose output (F=very little output, T=lots of output)}

\item{esrg}{location of e-surge program (eg., "C:/Program Files/E-SURGE 2.1.0/E_SURGE_V2_1_0_M9x64.exe")}
}
\value{
returns nothing.  Output for the model is saved in 2 files (*.out, *.xls).
}
\description{
Opens E-Surge window and automates the entry of the results filename (rslt),
input data filename, number of groups,...
}
\examples{
rslt="dipper.mod"; data="ed.rh";
ngrps=1; nicov=0; nstates=2; nevents=2; nage=1
f=c(list.files('.','^pi.+'),list.files('.','.+.mod$'),'tabdev.mat','shortcuts.out');
 #  assuming new analysis, so delete old model and/or output files
  supressWarnings(i=file.remove(f))
v=esrg_open(rslt,data,ngrps,nicov,nstates,nevents,nage,'noinit')
v=esrg_run('i','t','firste+nexte.t')     #   run model phi(t)p(t)    (init is ignored for this model.)
esrg_close()

\dontrun{
v=esrg_open(rslt,data,ngrps,nicov,nstates,nevents,nage,'noinit')
}

}
\seealso{
\code{\link{esrg_run}}, \code{\link{esrg_close}}
}

